refactor(netctl): split per-applet parsers/help into own files (#765)
Move each netctl applet's argument parser/planner and help (usage,
description, examples) into its own file (brctl.go, ifenslave.go,
tunctl.go, vconfig.go, zcip.go, nbd-client.go), leaving netctl.go as the
shared Command type, constructors, dispatch, and the shared
help()/usage() wiring routed through command.PlanGate. Behavior,
planned-action output, and gated error messages are unchanged.
